requirements.txt
# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=13.0.0  # For Parquet support

# Kafka
kafka-python>=2.0.2

# Monitoring
prometheus-client>=0.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data validation
great-expectations>=0.17.0

# Visualization (optional for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Elasticsearch client
elasticsearch>=8.9.0

# Time series processing
scipy>=1.11.0
statsmodels>=0.14.0
